Update version CRUD Orders

+++ WebAPI/API/ProductCategoriesAPIController.cs
@@ -14,6 +14,23 @@ namespace WebAPI.API
     {
         private readonly WebAPIContext db = new WebAPIContext();
         [HttpGet]
+        public IHttpActionResult ProductCategories()
+        {
+            var result = db.ProductCategories.OrderBy(i => i.Id)
+                .Select(s => new ProductCategoryViewModel
+                {
+                    Id = s.Id,
+                    CategoryCode = s.CategoryCode,
+                    CategoryName = s.CategoryName,
+                    CreateDate = s.CreateDate
+                });
+            return Ok(new
+            {
+                data = result,
+            });
+        }
+
+        [HttpGet]
         public IHttpActionResult ProductCategories(int skip, int take)
         {
             var result = db.ProductCategories.OrderBy(i => i.Id)
@@ -36,15 +53,14 @@ namespace WebAPI.API
         [HttpGet]
         public IHttpActionResult ProductCategory(int Id)
         {
-            var result = db.ProductCategories.Find(Id);
-            if (result != null)
+            var result = db.ProductCategories.Where(i => i.Id == Id).Select(s => new ProductCategoryViewModel
             {
-                return Ok(new { data = result });
-            }
-            else
-            {
-                return BadRequest();
-            }
+                Id = s.Id,
+                CategoryName = s.CategoryName,
+                CategoryCode = s.CategoryCode,
+                CreateDate = s.CreateDate,
+            }).SingleOrDefault();
+            return Ok(new { data = result });
 
         }
         [HttpPost]
@@ -76,9 +92,18 @@ namespace WebAPI.API
             var category = db.ProductCategories.Find(Id);
             if (category != null)
             {
-                db.ProductCategories.Remove(category);
-                db.SaveChanges();
-                return Ok();
+                var products = db.Products.Where(i => i.CategoryId == Id).ToList();
+                if (products.Count == 0)
+                {
+                    db.ProductCategories.Remove(category);
+                    db.SaveChanges();
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest();
+                }
+               
             }
             else
             {
@@ -91,9 +116,16 @@ namespace WebAPI.API
         [HttpPut]
         public IHttpActionResult ProductCategory(ProductCategoryViewModel model)
         {
+            ProductCategory category = new ProductCategory
+            {
+                Id=model.Id,
+                CategoryCode = model.CategoryCode,
+                CategoryName = model.CategoryName,
+                CreateDate=model.CreateDate,
+            };
             try
             {
-                db.Entry(model).State = EntityState.Modified;
+                db.Entry(category).State = EntityState.Modified;
                 db.SaveChanges();
             }
             catch
